Strengthen graph-slice ordering tests
Assert that graph-slice responses echo requested detail constraints and
that repeated requests preserve symbol ordering by stable symbol ID.

Move invalid-argument coverage into its own test module so the parent
module remains below the enforced line limit.
